chore: refatorando componente estrela
