+++ src/components/Estrelas.js
@@ -1,8 +1,6 @@
 import React, { useState } from "react";
-import { Image, StyleSheet, TouchableOpacity, View } from "react-native";
-
-import estrela from "../assets/estrela.png";
-import estrelaCinza from "../assets/estrelaCinza.png";
+import { StyleSheet, View } from "react-native";
+import Estrela from "./Estrela";
 
 export default function Estrelas({
     quantidade: quantidadeAntiga,
@@ -10,23 +8,18 @@ export default function Estrelas({
     grande = false,
 }) {
     const [ quantidade, setQuantidade ] = useState(quantidadeAntiga);
-    const estilos = estilosFuncao(grande);
-
-    const getImage = (index) => {
-        return index < quantidade ? estrela : estrelaCinza;
-    }
 
     const RenderEstrelas = () => {
         const listaEstrelas = [];
         for (let i = 0; i < 5; i++) {
             listaEstrelas.push(
-                <TouchableOpacity
+                <Estrela
                     key={i}
                     onPress={() => setQuantidade(i + 1)}
-                    disabled={!editavel}
-                    >
-                    <Image source={getImage(i)} style={estilos.estrela} />
-                </TouchableOpacity>
+                    desabilitada={!editavel}
+                    preenchida={i < quantidade}
+                    grande={grande}
+                />
             );
         }
         return listaEstrelas;
@@ -34,16 +27,11 @@ export default function Estrelas({
     
     return <View style={estilos.estrelas}>
         <RenderEstrelas />
-    </ View>
+    </View>
 }
 
-const estilosFuncao = (grande) => StyleSheet.create({
+const estilos = StyleSheet.create({
     estrelas: {
         flexDirection: 'row',
-    },
-    estrela: {
-        width: grande ? 36 : 12,
-        height: grande ? 36 : 12,
-        marginRight: 2,
     },
 });
